src/backend/requirements.txt
django==6.0a1
psycopg2-binary==2.9.10
fastkml==1.1.0
lxml==6.0.2
geojson==3.2.0
pydantic==2.11.9
markdownify==1.2.0
geoip2==5.1.0
python-magic==0.4.27
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
requests==2.31.0
pyyaml==6.0.1
pillow==12.0.0
whitenoise==6.11.0
django-allauth==65.13.0